requirements.txt
# Memo Filler Service - Dependencies
# For Fairbridge Deal Memo Generator

# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Template engine (Jinja2 for Word documents)
docxtpl==0.16.7
python-docx==1.1.0

# AWS S3 / DigitalOcean Spaces
boto3==1.34.14
botocore==1.34.14

# Image processing
Pillow==10.2.0
